Initial commit with program "finished"
Coursera beginning game programming

Assignment Description

Imagine you’re working on developing a game similar to Scorched Earth, a
2D tank-fighting game where tanks fire shells hoping to destroy the
enemy tanks. One of the challenges of the game is that the player has to
predict where the shell will land after following a parabolic
trajectory. While you may not have taken physics, the equations are
straightforward and you should be able to develop a solution that will
accurately calculate the shell path given a starting shell speed and
trajectory (angle of fire).